Move files from /redirects/engine directory to /redirects. Rewrite redirectFiles test.
