+++ domains/redirects/.gitignore
