+++ domains/redirects/RedirectHttp/function.json
