--- domains/redirects/engine/package.json
@@ -1,13 +1,0 @@
-{
-  "name": "redirect-engine",
-  "version": "1.0.0",
-  "description": "Subdomains redirect engine for ZHP.pl",
-  "scripts": {
-    "test": "jest"
-  },
-  "author": "Stanisław Kowański <skowanski@f4dev.me>",
-  "devDependencies": {
-    "jest": "^26.6.3",
-    "jest-json-schema": "^5.0.0"
-  }
-}
