+++ domains/redirects/host.json
