+++ domains/redirects/jest.config.js
