+++ domains/redirects/package.json
@@ -1,0 +1,13 @@
+{
+    "name": "redirect-engine",
+    "version": "1.0.0",
+    "description": "Subdomains redirect engine for ZHP.pl",
+    "scripts": {
+        "test": "node_modules/.bin/jest"
+    },
+    "author": "Stanisław Kowański <skowanski@f4dev.me>",
+    "devDependencies": {
+        "jest": "^26.6.3",
+        "ajv": "^8.4.0"
+    }
+}
